Create public-standard-pyFirmata
'Standard Firmata' is a public firmware designed for the communication between an Arduino board and external software like Python. 'pyFirmata' gives Python the command of the Arduino board for sending commands to its pins and reading sensor data and actuating devices such as LEDs, motors, and servos. In order for this communication to occur, the firmware 'Standard Firmata' must be uploaded to the Arduino-Issuing commands is easy for interfacing with Arduino boards using for various projects.