controller.py
The "controller.py" file therefore represents the intermediary between the Python script and the Arduino board to control LEDs depending on the gestures of the hand. It uses the 'pyFirmata' library to communicate with the Arduino on a specified COM port. A handful of digital pins on the Arduino board are outlined to which LEDs are connected, and these pins are used to control the LEDs. The led function takes a list corresponding to the state of the fingers (1 for up, 0 for down) and sets the corresponding LEDs on or off according to the state of the fingers. Thus, the system is capable of multiple LED control based on the gestures made by hand and is fully suitable for gesture-based control applications. The technique provides flexibility and must be easily extendable for the controlling of further hardware or sensors connected to the Arduino.