main-python-file.py
The primary file for executing Python lets one control a system powered by Arduino with a webcam's gesture . 'cvzone.HandTrackingModule' plays a vital role by determining how many fingers were raised based on hand detection. The video capturing comes through the cv2 library. The webcam image has been flipped for a mirror effect. When a hand is detected, it identifies the purpose of each finger raised and sends that count to the led() function in controller.py to handle LEDs depending on how many fingers are raised, directly bypassing the onboard controller. Continuous feedback to the users about finger counts is given in real-time. Different combinations such as "Thumbs up" and "Index and Middle Finger up" would be identified, with feedback given on the screen. This will be looping until the user presses the 'k' key, which will thereby release video capturing and exit the program. This creates continuous and interactive control for the users based on very simple hand gestures through Arduino.